+++ src/classes/Gantry/Component/Twig/TokenParser/TokenParserAssets.php
@@ -11,7 +11,9 @@
  * Gantry Framework code that extends GPL code is considered GNU/GPLv2 and later
  */
 
-namespace Gantry\Component\Twig;
+namespace Gantry\Component\Twig\TokenParser;
+
+use Gantry\Component\Twig\Node\TwigNodeScripts;
 
 /**
  * Adds javascript / style assets to head/footer/custom location.